fix(gui): UX improvements
