+++ adapter/src/main/java/fr/delivrooom/adapter/in/javafxgui/Sidebar.java
@@ -1,12 +1,21 @@
 package fr.delivrooom.adapter.in.javafxgui;
 
+import fr.delivrooom.adapter.in.javafxgui.command.AddDeliveryCommand;
+import fr.delivrooom.adapter.in.javafxgui.command.RemoveDeliveryCommand;
+import fr.delivrooom.adapter.in.javafxgui.controller.AppController;
+import fr.delivrooom.application.model.Delivery;
+import fr.delivrooom.application.model.Intersection;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.layout.VBox;
 
 public class Sidebar extends VBox {
-    public Sidebar() {
+
+    private final AppController controller;
+
+    public Sidebar(AppController controller) {
         super();
+        this.controller = controller;
 
         setAlignment(javafx.geometry.Pos.CENTER);
         setPadding(new javafx.geometry.Insets(10));
@@ -19,7 +28,27 @@ public class Sidebar extends VBox {
             label.setText("Name is: " + name);
         });
 
-        getChildren().addAll(label, button);
 
+        Button addDeliveryBtn = new Button("Add Delivery");
+        addDeliveryBtn.setOnAction(e -> handleAddDelivery());
+
+        Button removeDeliveryBtn = new Button("Remove Delivery");
+        removeDeliveryBtn.setOnAction(e -> handleRemoveDelivery());
+
+        getChildren().addAll(label, button, addDeliveryBtn, removeDeliveryBtn);
+    }
+
+    private void handleAddDelivery() {
+        Intersection takeoutIntersection = new Intersection(99, 45.550404, 4.8744674);
+        Intersection deliveryIntersection = new Intersection(100, 45.770404, 4.8744674);
+        Delivery addedDelivery = new Delivery(takeoutIntersection, deliveryIntersection, 5, 5);
+        AddDeliveryCommand addDeliveryCommand = new AddDeliveryCommand(controller, addedDelivery);
+        controller.getCommandManager().executeCommand(addDeliveryCommand);
+
+    }
+
+    private void handleRemoveDelivery() {
+        RemoveDeliveryCommand removeDeliveryCommand = new RemoveDeliveryCommand(controller, controller.getDeliveriesDemand().getDeliveryByIds(99, 100));
+        controller.getCommandManager().executeCommand(removeDeliveryCommand);
     }
 }
